Problem with BigInt

+++ web/src/main/java/com/infoshareacademy/wojownicy/service/BookListService.java
@@ -40,7 +40,7 @@ public class BookListService {
     return pagesMap;
   }
 
-  public List<BookDto> partOfBooks(int from, long kind) {
+  public List<BookDto> partOfBooks(int from, BigInteger kind) {
    List<Book> books = bookDaoBean.getPartOfBooks(from);
    List<BookDto> bookDtoList = new ArrayList<>();
    for (Book book : books){
@@ -49,7 +49,7 @@ public class BookListService {
   return  bookDtoList;
   }
 
-  public List<BookDto> partOfAudioBooks(int from, long kind){
+  public List<BookDto> partOfAudioBooks(int from, BigInteger kind){
     List<BookDto> bookDtoList = new ArrayList<>();
 
     List<Book> books = bookDaoBean.getPartOfAudioBooks(from, kind);
